spectai-overlay-v2/requirements.txt
PyQt6>=6.6.0
supabase>=2.0.0
python-dotenv>=1.0.0
